# restAPI/src/main/swagger/openapi.yaml
openapi: 3.0.0
info:
  description: This is the API for the UI's access to data storage
  version: 1.0.0
  title: AIDA Store REST API
servers:
  - url: "{protocol}://{host}:{port}/{basePath}"
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: "8008"
      basePath:
        default: api
...
paths:
...
  /relations:
    get:
      tags:
        - Relation
      summary: Return relations matching provided query inputs
      description: Return relations by graph and filter by multiple query inputs.
      operationId: relations
      parameters:
        - $ref: "#/components/parameters/graphParam"
        - $ref: "#/components/parameters/argumentsParam"
        - $ref: "#/components/parameters/typeFilter"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CachedEvent"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
  /relations/types:
    get:
      tags:
        - Relation
      summary: Return relation types matching provided input
      description: Return relation types by graph and filter by specified input
      operationId: relationTypes
      parameters:
        - $ref: "#/components/parameters/graphParam"
        - name: type
          in: query
          description: Part or all of relation type
          required: false
          schema:
            type: string
          example: Physical.LocatedNear
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "400":
          description: Invalid parameters
        "404":
          description: API not found
  /relations/arguments:
    get:
      tags:
        - Relation
      summary: Return relation argument names matching provided input
      description: Return relation argument names by graph and filter by specified input
      operationId: relationNames
      parameters:
        - $ref: "#/components/parameters/graphParam"
        - $ref: "#/components/parameters/argumentParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "400":
          description: Invalid parameters
        "404":
          description: API not found
  /relation:
    get:
      tags:
        - Relation
      summary: Return a single relation
      description: Return relation by id
      operationId: relation
      parameters:
        - name: id
          in: query
          description: ID of relation to retrieve
          required: true
          schema:
            type: string
          example: http://www.lti.cs.cmu.edu/aida/opera/corpora/eval/cluster-relation-instance-JC002YEM4-r202010162240-43
        - $ref: "#/components/parameters/optionalGraphParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
  /relation/element:
    get:
      tags:
        - Relation
      summary: Return a single relation element
      description: Return relation element by id
      operationId: relationElement
      parameters:
        - name: id
          in: query
          description: ID of relation element to retrieve
          required: true
          schema:
            type: string
          example: http://www.lti.cs.cmu.edu/aida/opera/corpora/eval/relation-instance-JC002YEM4-r202010162240-43
        - $ref: "#/components/parameters/optionalGraphParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Element"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
...
components:
  schemas:
...
    CachedEvent:
      type: object
      properties:
        cluster:
          type: string
        handle:
          type: string
        prototype:
          type: string
        types:
          type: array
          items:
            type: string
        names:
          type: array
          items:
            type: string
        argumentCount:
          type: integer
    Event:
      type: object
      properties:
        cluster:
          type: string
        handle:
          type: string
        category:
          type: string
        prototype:
          $ref: "#/components/schemas/Member"
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
    Member:
      allOf:
        - $ref: "#/components/schemas/JustifiedNode"
        - type: object
          properties:
            arguments:
              type: array
              items:
                $ref: "#/components/schemas/Role"
    Element:
      allOf:
        - $ref: "#/components/schemas/Member"
        - type: object
          properties:
            clusters:
              type: array
              items:
                $ref: "#/components/schemas/JustifiedNode"
    Role:
      type: object
      properties:
        role:
          type: string
        fillers:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
    JustifiedNode:
      type: object
      discriminator:
        propertyName: objectType
      properties:
        id:
          type: string
        handle:
          type: string
        objectType:
          type: string
          enum: ["Entity", "Member", "Element", "EntityMember", "EntityElement", "JustifiedNode"]
        types:
          type: array
          items:
            type: string
        docs:
          type: array
          items:
            $ref: "#/components/schemas/DocObject"
    Entity:
      allOf:
        - $ref: "#/components/schemas/JustifiedNode"
        - type: object
          properties:
            names:
              type: array
              items:
                type: string
            category:
              type: string
            clusterIds:
              type: array
              items:
                type: string
...
    DocObject:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        contentDate:
          type: string
        downloadDate:
          type: string
...
  parameters:
    graphParam:
      name: graph
      in: query
      description: Which graph to query in
      required: true
      schema:
        type: string
      example: https://www.nextcentury.com/TA2/OPERA-20210215/OPERA-20210117
    optionalGraphParam:
      name: graph
      in: query
      description: Which graph to query in
      schema:
        type: string
      example: https://www.nextcentury.com/TA2/OPERA-20210215/OPERA-20210117
...
    argumentParam:
      name: argument
      in: query
      description: Part or all of argument name
      required: false
      schema:
        type: string
      example: Chavez
    limitParam:
      name: dataLimit
      in: query
      description: The limit of the number of data returned.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 500
    argumentsParam:
      name: arguments
      in: query
      description: Participant names to filter by
      required: false
      schema:
        type: array
        items:
          type: string
          example: Chavez
          uniqueItems: true
      example:
        - Chavez
    typeFilter:
      name: types
      in: query
      description: The types to filter by
      required: false
      schema:
        type: array
        items:
          type: string
          example: https://raw.githubusercontent.com/NextCenturyCorporation/AIDA-Interchange-Format/master/java/src/main/resources/com/ncc/aif/ontologies/LDCOntologyM36#OrganizationAffiliation.Leadership
          uniqueItems: true
      example:
        - https://raw.githubusercontent.com/NextCenturyCorporation/AIDA-Interchange-Format/master/java/src/main/resources/com/ncc/aif/ontologies/LDCOntologyM36#Life.Die
        - https://raw.githubusercontent.com/NextCenturyCorporation/AIDA-Interchange-Format/master/java/src/main/resources/com/ncc/aif/ontologies/LDCOntologyM36#OrganizationAffiliation.Leadership
...